requirements.txt
2to3==1.0
alembic==1.11.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.6.0
Babel==2.10.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.10.0
bleach==5.0.0
blinker==1.6.2
caer==2.0.8
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.5
Django==4.0.6
entrypoints==0.4
executing==0.8.3
face-recognition-models==0.3.0
fastjsonschema==2.15.3
filelock==3.7.1
flask==2.3.3
Flask-Migrate==4.0.4
flask-sqlalchemy==3.0.5
greenlet==2.0.2
idna==3.3
importlib-metadata==6.8.0
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
joblib==1.0.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.3.1
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.20.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
MouseInfo==0.1.3
mypy==0.930
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
nltk==3.6.2
notebook==6.4.10
numpy==1.21.0
opencv-contrib-python==4.5.5.62
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.68
packaging==21.0
pandas==1.3.1
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
platformdirs==2.5.2
pluggy==0.13.1
prometheus-client==0.14.1
prompt-toolkit==3.0.19
protobuf==3.19.4
psutil==5.9.0
psycopg2==2.9.7
pure-eval==0.2.2
py==1.10.0
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.9.0
pymongo==4.2.0
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
pyrsistent==0.18.1
PyScreeze==0.1.28
pytest==6.2.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.4
pytz==2021.1
pywhatkit==5.3
pywin32==301
pywinpty==2.0.5
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
regex==2021.7.6
requests==2.27.1
scikit-learn==0.24.2
scipy==1.7.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
SQLAlchemy==2.0.20
sqlparse==0.4.2
stack-data==0.2.0
terminado==0.13.3
threadpoolctl==2.2.0
tinycss2==1.1.1
tk==0.1.0
tkcalendar==1.6.1
toml==0.10.2
tomli==2.0.0
tornado==6.1
tqdm==4.61.2
traitlets==5.1.1
typing-extensions==4.7.1
tzdata==2022.1
urllib3==1.26.8
virtualenv==20.15.1
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==2.3.7
widgetsnbextension==3.6.0
wikipedia==1.4.0
zipp==3.16.2
